+++ CHANGES.rst
@@ -43,6 +43,8 @@ Changelog
   moreover, display the `WorkflowState` viewlet on `person view` and `held_position view`
 - Fixed `showAddAnnex` and `showAddAnnexDecision` in `@@categorized-annexes`, rely on the `content_category` field vocabulary
 
+- Fix MeetingUser migration when no gender setted
+
 4.1.22.1 (2020-04-24)
 ---------------------